code: project restructure
