+++ lib/lrc.dart
@@ -1,4 +1,11 @@
 /// The main library for support for LRCs.
 library lrc;
 
-export 'src/lrc_main.dart';
+part 'src/core/enums.dart';
+part 'src/core/extensions.dart';
+part 'src/models/lrc.dart';
+part 'src/models/lrc_line.dart';
+part 'src/models/lrc_line_part.dart';
+part 'src/models/lrc_stream.dart';
+part 'src/parsers/lrc_parser.dart';
+part 'src/parsers/multi_timestamp_parser.dart';
