+++ lib/src/models/lrc_line_part.dart
@@ -1,0 +1,19 @@
+part of lrc;
+
+class LrcLinePart {
+  final Duration timestamp;
+  final String lyrics;
+
+  const LrcLinePart({
+    required this.timestamp,
+    required this.lyrics,
+  });
+
+  @override
+  String toString() {
+    return '''
+      Timestamp: '$timestamp'
+      Lyrics: '$lyrics'
+    ''';
+  }
+}
